release: v1.1.0 for eslint-config

+++ apps/eslint-config/src/default.ts
@@ -37,7 +37,7 @@ const config: Linter.Config = {
     rules: {
         quotes: [
             2,
-            "single",
+            "double",
             {
                 avoidEscape: true,
             },
